unnecessary files have been deleted
