+++ client/app/(pages)/signup/SignUp.scss
@@ -8,5 +8,6 @@
   &__title {
     font-weight: 400;
     margin-bottom: 20px;
+    text-align: center;
   }
 }
