+++ client/app/(pages)/wallet/Wallet.scss
@@ -10,5 +10,6 @@
   &__title {
     font-weight: 400;
     margin-bottom: 20px;
+    text-align: center;
   }
 }
