+++ client/app/components/DiscoverNftSection/DiscoverNftSection.tsx
@@ -3,41 +3,7 @@ import "./DiscoverNftSection.scss";
 import eyeIcon from "../../../public/eyeIcon.svg";
 import { DiscoveryCard } from "./DiscoveryCard";
 import { BASE_URL } from "@/app/utils/endpoint";
-import { IAsset } from "@/app/types";
-
-import { gql } from "@apollo/client";
-import { getClient } from "@/app/libs/client";
-
-export async function getAssets(): Promise<IAsset[]> {
-  const query = gql`
-    query {
-      assets{
-    _id
-    imgUrl
-    assetName
-    price
-    highestBid
-    createdBy{
-      nickName
-      avatarUrl
-      _id
-    }
-  }
-    }
-  `;
-
-  const assets = await getClient()
-    .query<{ assets: IAsset[] }>({
-      query,
-    })
-    .then((res) => res.data)
-    .catch(e => {
-      console.log('ERROR home-discovery-route =>', e);
-      return {assets: []}
-    })
-
-    return assets.assets;
-};
+import { getAssets } from "./helper";
 
 export default async function DiscoverNftSection() {
   const cards = (await getAssets()).slice(0, 3);
