+++ client/app/components/Footer/Footer.tsx
@@ -6,68 +6,14 @@ import { SubscribeInput } from "./SubscribeInput";
 import youtubeIcon from "../../../public/youtubeLogo.svg";
 import twitterIcon from "../../../public/twitterLogo.svg";
 import instagramIcon from "../../../public/instagramLogo.svg";
-import { ISocialLink, ISudscribe } from "@/app/types";
 import Image from "next/image";
-
-import { gql } from "@apollo/client";
-import { getClient } from "@/app/libs/client";
+import { getSocialLinks, getSubscribe } from "./helper";
 
 const socialLinks: { [key: string]: any } = {
   youtube: youtubeIcon as any,
   twitter: twitterIcon as any,
   instagram: instagramIcon as any,
 };
-
-export async function getSubscribe(): Promise<ISudscribe> {
-  const query = gql`
-    query {
-      subscribe {
-        imgUrl
-        title
-        description
-      }
-    }
-  `;
-
-  const subscribe = await getClient()
-    .query<{ subscribe: ISudscribe[] }>({
-      query,
-    })
-    .then((res) => res.data.subscribe[0])
-    .catch((e) => {
-      console.log("ERROR home-subscribe-route =>", e);
-      return {
-        imgUrl: "",
-        title: "",
-        description: "",
-      };
-    });
-
-  return subscribe;
-}
-
-export async function getSocialLinks(): Promise<ISocialLink[]> {
-  const query = gql`
-    query {
-      socialLinks {
-        name
-        url
-      }
-    }
-  `;
-
-  const socialLinks = await getClient()
-    .query<{ socialLinks: ISocialLink[] }>({
-      query,
-    })
-    .then((res) => res.data.socialLinks)
-    .catch((e) => {
-      console.log("ERROR footer-route =>", e);
-      return [];
-    });
-
-  return socialLinks;
-}
 
 export default async function Footer() {
   const subscribe = await getSubscribe();
