+++ client/app/not-found.tsx
@@ -1,5 +1,4 @@
 import Link from 'next/link'
-import { Button } from './components/Button';
  
 export default function Error() {
